fix(gui): handle invalid file name edge case on Windows
File names like "c:blah" would break the usage of the dialogue,
displaying a path like "c://blah" instead.

Apparently `QDir::filePath` checks if the passed file name is an
absolute file name using `!QFileInfo::isRelative`, which is the case if
the second character in the string is a colon -- probably indicating a
drive letter.

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>