fix(download): allow highly compressed responses up to a known file size
Qt 6.2+'s archive bomb check is cool and all, but it's really annoying
when trying to retrieve files that compress really well over the wire.
As we already know the expected file size before the request we can
basically tell Qt to allow decompression of responses up to the expected
file size and a bit to spare.

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>