chore: add test for response decompression threshold check
Unfortunately faking this with the usual methods won't do the trick:
setting an override in our FakeQNAM returning a faked reply completely
bypasses Qt's internal routines for QNetworkReply.
Therefore: spin up a temporary local HTTP server for the time of the
test and do real HTTP requests.

We could try to explore how useful QHttpServer would be in other tests
in the future as well, it might enable us for a cleaner way of setting
up a fake/mock server for unit tests.  For now this specific test can be
seen as the initial exploration on how it could be used.

Caveat: this test is only available if the QHttpServer component is
present, chances are it may not run everywhere.

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>